# src/main/java/com/assessment/demo/vo/UserChangePasswordModel.java
package com.assessment.demo.vo;

public class UserChangePasswordModel {
    String password;

    public String getPassword() {
        return password;
    }

    public void setPassword(String password) {
        this.password = password;
    }
    
}