# swagger-autogenerated_JSON.json
{
  "swagger": "2.0",
  "info": {
    "description": "Spring boot assessment task by Maria",
    "title": "Assessment Task"
  },
  "host": "localhost:8080",
  "basePath": "/",
  "tags": [{ "name": "Available APIs", "description": "User Controller" }],
  "paths": {
    "/users": {
      "get": {
        "tags": ["Available APIs"],
        "summary": "List all users",
        "operationId": "getAllUsersUsingGET",
        "produces": ["*/*"],
        "responses": {
          "200": {
            "description": "Fetched all users successfully",
            "schema": {
              "type": "array",
              "items": { "$ref": "#/definitions/UserDisplayInfo" }
            }
          },
          "400": { "description": "Failed due to bad input" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Not Found" }
        }
      },
      "post": {
        "tags": ["Available APIs"],
        "summary": "create new user",
        "operationId": "createUserUsingPOST",
        "consumes": ["application/json"],
        "produces": ["*/*"],
        "parameters": [
          {
            "in": "body",
            "name": "user",
            "description": "user",
            "required": true,
            "schema": { "$ref": "#/definitions/UserInfo" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": { "$ref": "#/definitions/ResponseEntity" }
          },
          "201": { "description": "Successfully created the user" },
          "400": { "description": "Failed due to bad input" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Not Found" }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": ["Available APIs"],
        "summary": "Get user details using id",
        "operationId": "getUserByIdUsingGET",
        "produces": ["*/*"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Fetched user details successfully",
            "schema": { "$ref": "#/definitions/UserDisplayInfo" }
          },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "404": { "description": "No user found with the given id" }
        }
      },
      "put": {
        "tags": ["Available APIs"],
        "summary": "Update user details",
        "operationId": "updateUserDetailsUsingPUT",
        "consumes": ["application/json"],
        "produces": ["*/*"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "userDisplayInfo",
            "description": "userDisplayInfo",
            "required": true,
            "schema": { "$ref": "#/definitions/UserDisplayInfo" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": { "$ref": "#/definitions/ResponseEntity" }
          },
          "201": { "description": "Created" },
          "204": { "description": "Updated Successfully" },
          "400": { "description": "Update failed due to bad input" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Unable to find user with the given id" }
        }
      },
      "delete": {
        "tags": ["Available APIs"],
        "summary": "Delete a user",
        "operationId": "deleteUserUsingDELETE",
        "produces": ["*/*"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted the user successfully",
            "schema": { "$ref": "#/definitions/ResponseEntity" }
          },
          "204": { "description": "No Content" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Unable to find user with the given id" }
        }
      },
      "patch": {
        "tags": ["Available APIs"],
        "summary": "Change password",
        "operationId": "changePasswordUsingPATCH",
        "consumes": ["application/json"],
        "produces": ["*/*"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "userNewPassword",
            "description": "userNewPassword",
            "required": true,
            "schema": { "$ref": "#/definitions/UserChangePasswordModel" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": { "$ref": "#/definitions/ResponseEntity" }
          },
          "204": { "description": "Updated Password Successfully" },
          "400": { "description": "Update failed due to bad input" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Unable to find user with the given id" }
        }
      }
    }
  },
  "definitions": {
    "ResponseEntity": {
      "type": "object",
      "properties": {
        "body": { "type": "object" },
        "statusCode": {
          "type": "string",
          "enum": [
            "ACCEPTED",
            "ALREADY_REPORTED",
            "BAD_GATEWAY",
            "BAD_REQUEST",
            "BANDWIDTH_LIMIT_EXCEEDED",
            "CHECKPOINT",
            "CONFLICT",
            "CONTINUE",
            "CREATED",
            "DESTINATION_LOCKED",
            "EXPECTATION_FAILED",
            "FAILED_DEPENDENCY",
            "FORBIDDEN",
            "FOUND",
            "GATEWAY_TIMEOUT",
            "GONE",
            "HTTP_VERSION_NOT_SUPPORTED",
            "IM_USED",
            "INSUFFICIENT_SPACE_ON_RESOURCE",
            "INSUFFICIENT_STORAGE",
            "INTERNAL_SERVER_ERROR",
            "I_AM_A_TEAPOT",
            "LENGTH_REQUIRED",
            "LOCKED",
            "LOOP_DETECTED",
            "METHOD_FAILURE",
            "METHOD_NOT_ALLOWED",
            "MOVED_PERMANENTLY",
            "MOVED_TEMPORARILY",
            "MULTIPLE_CHOICES",
            "MULTI_STATUS",
            "NETWORK_AUTHENTICATION_REQUIRED",
            "NON_AUTHORITATIVE_INFORMATION",
            "NOT_ACCEPTABLE",
            "NOT_EXTENDED",
            "NOT_FOUND",
            "NOT_IMPLEMENTED",
            "NOT_MODIFIED",
            "NO_CONTENT",
            "OK",
            "PARTIAL_CONTENT",
            "PAYLOAD_TOO_LARGE",
            "PAYMENT_REQUIRED",
            "PERMANENT_REDIRECT",
            "PRECONDITION_FAILED",
            "PRECONDITION_REQUIRED",
            "PROCESSING",
            "PROXY_AUTHENTICATION_REQUIRED",
            "REQUESTED_RANGE_NOT_SATISFIABLE",
            "REQUEST_ENTITY_TOO_LARGE",
            "REQUEST_HEADER_FIELDS_TOO_LARGE",
            "REQUEST_TIMEOUT",
            "REQUEST_URI_TOO_LONG",
            "RESET_CONTENT",
            "SEE_OTHER",
            "SERVICE_UNAVAILABLE",
            "SWITCHING_PROTOCOLS",
            "TEMPORARY_REDIRECT",
            "TOO_EARLY",
            "TOO_MANY_REQUESTS",
            "UNAUTHORIZED",
            "UNAVAILABLE_FOR_LEGAL_REASONS",
            "UNPROCESSABLE_ENTITY",
            "UNSUPPORTED_MEDIA_TYPE",
            "UPGRADE_REQUIRED",
            "URI_TOO_LONG",
            "USE_PROXY",
            "VARIANT_ALSO_NEGOTIATES"
          ]
        },
        "statusCodeValue": { "type": "integer", "format": "int32" }
      },
      "title": "ResponseEntity"
    },
    "UserChangePasswordModel": {
      "type": "object",
      "properties": { "password": { "type": "string" } },
      "title": "UserChangePasswordModel"
    },
    "UserDisplayInfo": {
      "type": "object",
      "properties": {
        "emailId": { "type": "string" },
        "id": { "type": "integer", "format": "int64" },
        "name": { "type": "string" }
      },
      "title": "UserDisplayInfo"
    },
    "UserInfo": {
      "type": "object",
      "properties": {
        "emailId": { "type": "string" },
        "id": { "type": "integer", "format": "int64" },
        "name": { "type": "string" },
        "password": { "type": "string" }
      },
      "title": "UserInfo"
    }
  }
}